Add tests: qdb_deque_pop_back, qdb_deque_pop_front, qdb_int_get, qdb_int_put and qdb_int_remove

+++ bench/tests/qdb/qdb_test_template.hpp
@@ -37,7 +37,7 @@ public:
 protected:
     utils::qdb_wrapper _qdb;
 
-    std::string get_alias(long long iteration) const
+    std::string get_alias(unsigned long iteration) const
     {
         return _alias_prefix + std::to_string(iteration);
     }
